Fix some uses of SimTK::Xml in preparation for Xml being promoted to namespace. Fix miscellaneous warnings caught by VC++.

+++ OpenSim/Common/DataTable.h
@@ -258,7 +258,7 @@ public:
     RowVectorView getRowAtIndex(size_t index) const {
         OPENSIM_THROW_IF(isRowIndexOutOfRange(index),
                          RowIndexOutOfRange, index, 0, _indData.size())
-        return _depData.row(index);
+        return _depData.row((int)index);
     }
 
     /** Get row corresponding to the given entry in the independent column.   
@@ -271,7 +271,7 @@ public:
         OPENSIM_THROW_IF(iter == _indData.cend(),
                          KeyNotFound, std::to_string(ind));
 
-        return _depData.row(std::distance(_indData.cbegin(), iter));
+        return _depData.row((int)std::distance(_indData.cbegin(), iter));
     }
 
     /** Update row at index.                                                  
@@ -280,7 +280,7 @@ public:
     RowVectorView updRowAtIndex(size_t index) {
         OPENSIM_THROW_IF(isRowIndexOutOfRange(index),
                          RowIndexOutOfRange, index, 0, _indData.size());
-        return _depData.updRow(index);
+        return _depData.updRow((int)index);
     }
 
     /** Update row corresponding to the given entry in the independent column.
@@ -293,7 +293,7 @@ public:
         OPENSIM_THROW_IF(iter == _indData.cend(),
                          KeyNotFound, std::to_string(ind));
 
-        return _depData.updRow(std::distance(_indData.cbegin(), iter));
+        return _depData.updRow((int)std::distance(_indData.cbegin(), iter));
     }
 
     /** Get independent column.                                               */
@@ -308,7 +308,7 @@ public:
         OPENSIM_THROW_IF(isColumnIndexOutOfRange(index),
                          ColumnIndexOutOfRange, index, 0,
                          static_cast<size_t>(_depData.ncol()));
-        return _depData.col(index);
+        return _depData.col((int)index);
     }
 
     /** Set independent column at index.                                      
@@ -319,7 +319,7 @@ public:
     void setIndependentColumnAtIndex(size_t index, const ETX& value) {
         OPENSIM_THROW_IF(isRowIndexOutOfRange(index),
                          RowIndexOutOfRange, index, 0, _indData.size());
-        validateRow(index, value, _depData.row(index));
+        validateRow(index, value, _depData.row((int)index));
         _indData[index] = value;
     }
 
@@ -365,7 +365,8 @@ protected:
     void validateDependentsMetaData() const override {
         unsigned numCols{};
         try {
-            numCols = _dependentsMetaData.getValueArrayForKey("labels").size();
+            numCols = (unsigned)_dependentsMetaData
+                                        .getValueArrayForKey("labels").size();
         } catch (KeyNotFound&) {
             OPENSIM_THROW(MissingMetaData, "labels");
         }
